+++ packages/plugins/lazy-load/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-lazy-load
 
+## 2.0.0
+
 ## 1.1.0
 
 ### Minor Changes