+++ packages/plugins/selectable/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @splicetree/plugin-selectable
 
+## 2.0.0
+
 ## 1.1.0
 
 ## 1.0.0